+++ src/main/java/chess/WebUIChessApplication.java
@@ -4,7 +4,9 @@ import static spark.Spark.get;
 import static spark.Spark.post;
 import static spark.Spark.put;
 
-import chess.dao.GameDao;
+import chess.dao.PlayLogDao;
+import chess.dao.RoomDao;
+import chess.dao.UserDao;
 import chess.domain.board.Board;
 import chess.domain.board.Point;
 import chess.domain.board.Team;
@@ -12,11 +14,11 @@ import chess.domain.chessgame.ChessGame;
 import chess.domain.chessgame.ScoreBoard;
 import chess.domain.chessgame.Turn;
 import chess.domain.gamestate.GameState;
-import chess.dto.BoardWebDto;
-import chess.dto.GameStatusDto;
-import chess.dto.PointDto;
-import chess.dto.RoomDto;
-import chess.dto.RoomUsersDto;
+import chess.dto.web.BoardDto;
+import chess.dto.web.GameStatusDto;
+import chess.dto.web.PointDto;
+import chess.dto.web.RoomDto;
+import chess.dto.web.UsersInRoomDto;
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import java.sql.SQLException;
@@ -31,7 +33,10 @@ import spark.template.handlebars.HandlebarsTemplateEngine;
 public class WebUIChessApplication {
 
     private static final Gson GSON = new Gson();
-    private static final GameDao GAME_DAO = new GameDao();
+    private static final PlayLogDao PLAY_LOG_DAO = new PlayLogDao();
+    private static final RoomDao ROOM_DAO = new RoomDao();
+    private static final UserDao USER_DAO = new UserDao();
+
 
     public static void main(String[] args) {
 
@@ -39,16 +44,16 @@ public class WebUIChessApplication {
 
         get("/", (req, res) -> {
             Map<String, Object> model = new HashMap<>();
-            model.put("rooms", GAME_DAO.openedRooms());
+            model.put("rooms", ROOM_DAO.openedRooms());
             return render(model, "lobby.html");
         });
 
         post("/room", "application/json", (req, res) -> {
             try {
                 RoomDto newRoom = GSON.fromJson(req.body(), RoomDto.class);
-                GAME_DAO.insertUser(newRoom.getWhite());
-                GAME_DAO.insertUser(newRoom.getBlack());
-                String roomId = GAME_DAO.insertRoom(newRoom);
+                USER_DAO.insert(newRoom.getWhite());
+                USER_DAO.insert(newRoom.getBlack());
+                String roomId = ROOM_DAO.insert(newRoom);
                 JsonObject jsonObject = new JsonObject();
                 jsonObject.addProperty("result", "success");
                 jsonObject.addProperty("roomId", roomId);
@@ -62,12 +67,12 @@ public class WebUIChessApplication {
         });
 
         get("/room/:id", (req, res) -> {
-            BoardWebDto boardWebDto = GAME_DAO.latestBoard(req.params("id"));
-            RoomUsersDto roomUsersDto = GAME_DAO.roomUsers(req.params("id"));
+            BoardDto boardDto = PLAY_LOG_DAO.latestBoard(req.params("id"));
+            UsersInRoomDto usersInRoomDto = USER_DAO.usersInRoom(req.params("id"));
             Map<String, Object> model = new HashMap<>();
-            model.put("board", boardWebDto);
+            model.put("board", boardDto);
             model.put("roomId", req.params(":id"));
-            model.put("userInfo", roomUsersDto);
+            model.put("userInfo", usersInRoomDto);
             return render(model, "index.html");
         });
 
@@ -75,10 +80,10 @@ public class WebUIChessApplication {
             Board board = boardFromDb(req.params(":id"));
             ChessGame chessGame = chessGameFromDb(board, req.params(":id"));
             chessGame.start();
-            BoardWebDto boardWebDto = new BoardWebDto(board);
+            BoardDto boardDto = new BoardDto(board);
             GameStatusDto gameStatusDto = new GameStatusDto(chessGame);
-            GAME_DAO.insertBoardAndStatusDto(boardWebDto, gameStatusDto, req.params(":id"));
-            return GSON.toJson(boardWebDto);
+            PLAY_LOG_DAO.insert(boardDto, gameStatusDto, req.params(":id"));
+            return GSON.toJson(boardDto);
         });
 
         get("/room/:id/movablePoints/:point", "application/json", (req, res) -> {
@@ -94,7 +99,7 @@ public class WebUIChessApplication {
         });
 
         get("/room/:id/stat", "application/json", (req, res) -> {
-            return GSON.toJson(GAME_DAO.roomUsers(req.params("id")));
+            return GSON.toJson(USER_DAO.usersInRoom(req.params("id")));
         });
 
         put("/room/:id/move", "application/json", (req, res) -> {
@@ -105,13 +110,12 @@ public class WebUIChessApplication {
             Point source = Point.of(body.get("source"));
             Point destination = Point.of(body.get("destination"));
             chessGame.move(source, destination);
-            GAME_DAO
-                .insertBoardAndStatusDto(new BoardWebDto(board), new GameStatusDto(chessGame),
-                    roomId);
+            PLAY_LOG_DAO
+                .insert(new BoardDto(board), new GameStatusDto(chessGame), roomId);
             if (!chessGame.isOngoing() && chessGame.winner() != Team.NONE) {
-                GAME_DAO.addUserStat(roomId, chessGame.winner());
+                USER_DAO.updateStats(roomId, chessGame.winner());
             }
-            return GSON.toJson(new BoardWebDto(board));
+            return GSON.toJson(new BoardDto(board));
         });
 
         get("/room/:id/getGameStatus", "application/json", (req, res) -> {
@@ -126,26 +130,26 @@ public class WebUIChessApplication {
             Board board = boardFromDb(roomId);
             ChessGame chessGame = chessGameFromDb(board, roomId);
             chessGame.end();
-            BoardWebDto boardWebDto = new BoardWebDto(board);
+            BoardDto boardDto = new BoardDto(board);
             GameStatusDto gameStatusDto = new GameStatusDto(chessGame);
-            GAME_DAO.insertBoardAndStatusDto(boardWebDto, gameStatusDto, req.params(":id"));
+            PLAY_LOG_DAO.insert(boardDto, gameStatusDto, req.params(":id"));
             return GSON.toJson("success");
         });
 
         put("/room", (req, res) -> {
             Map<String, String> body = GSON.fromJson(req.body(), HashMap.class);
-            GAME_DAO.closeRoom(body.get("id"));
+            ROOM_DAO.close(body.get("id"));
             return GSON.toJson("success");
         });
     }
 
     private static Board boardFromDb(String roomId) throws SQLException {
-        return GAME_DAO.latestBoard(roomId).toEntity();
+        return PLAY_LOG_DAO.latestBoard(roomId).toEntity();
     }
 
     private static ChessGame chessGameFromDb(Board board, String roomId)
         throws SQLException {
-        GameStatusDto gameStatusDto = GAME_DAO.latestGameStatus(roomId);
+        GameStatusDto gameStatusDto = PLAY_LOG_DAO.latestGameStatus(roomId);
         Turn turn = gameStatusDto.toTurnEntity();
         GameState gameState = gameStatusDto.toGameStateEntity(board);
         return new ChessGame(turn, new ScoreBoard(board), gameState);
