+++ src/main/java/chess/dto/console/BoardDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.console;
 
 import chess.domain.board.Board;
 import chess.domain.board.Column;
