+++ src/main/java/chess/dto/web/GameStatusDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.web;
 
 import chess.domain.board.Board;
 import chess.domain.board.Team;
