+++ src/main/java/chess/dto/web/PieceDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.web;
 
 import chess.domain.board.Point;
 import chess.domain.board.SquareState;
