+++ src/main/java/chess/dto/web/RoomDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dto.web;
 
 public class RoomDto {
